+++ changelog.md
@@ -8,6 +8,7 @@
 - Extract value preparation from send_value
 - Differentiate between input and default parameter type
 - Fix asyncio-dangling-task (RUF006)
+- Add typed dict for light and siren
 
 # Version 2023.2.8 (2023-02-11)